+++ mock/login.json
@@ -4,6 +4,6 @@
     "data": {
       "userType": "团队长",
       "msg": "", 
-      "um": "admin"
+      "um": "guanguan"
     }
 }
