+++ src/component/login/index.jsx
@@ -1,0 +1,3 @@
+import Login from './Login';
+
+export default Login;
