+++ src/container/login/index.less
@@ -1,30 +1,3 @@
-@import "../../../style/theme/default/index.less";
-.main {
-  width: 368px;
-  margin: 0 auto;
-
-  :global {
-    .idoll-tabs .idoll-tabs-bar {
-      border-bottom: 0;
-      margin-bottom: 24px;
-      text-align: center;
-    }
-
-    .idoll-form-item {
-      margin-bottom: 24px;
-    }
-  }
-
-  .prefixIcon {
-    font-size: @font-size-base;
-    color: @disabled-color;
-  }
-
-  .submit {
-    width: 100%;
-    margin-top: 24px;
-  }
-}
 .loginpage {
   width: 100%;
   min-height: 200%;
